+++ DummyMap.h
@@ -1,8 +1,8 @@
 #pragma once
 #include <vector>
 
-extern std::vector<std::vector<int> > dummyMap;
+extern std::vector<std::vector<int16_t> > dummyMap;
 
 // s—ñ‚ğ“]’u‚·‚é
-std::vector<std::vector<int> > transporseMatrix(
-    const std::vector<std::vector<int> > &matrix);
+std::vector<std::vector<int16_t> > transporseMatrix(
+    const std::vector<std::vector<int16_t> > &matrix);
